[#12] Move files
